+++ mobile/android/android-components/components/feature/tabs/src/main/java/mozilla/components/feature/tabs/ext/TabSessionState.kt
@@ -15,5 +15,6 @@ internal fun TabSessionState.toTab() = Tab(
     content.icon,
     content.thumbnail,
     mediaSessionState?.playbackState,
-    mediaSessionState?.controller
+    mediaSessionState?.controller,
+    lastAccess
 )